Internal: move scrubber and time rendering to sub components
By moving the connected updates to scrubber and time display to sub
components we avoid re-rendering the broader playback control panel
during playback as the current time changes. Since the current time
may update at 60hz this saves us some rendering of non-essential components.
